reorganized dims and types
